looks like converting from a postfix string to a postfix tree works

+++ calculator/src/treeCommon.c
@@ -1,6 +1,7 @@
 #include "treeCommon.h"
 #include<stdlib.h>
 #include<string.h>
+#include<stdio.h>
 
 /*
 Private prototypes
@@ -11,8 +12,18 @@ Private prototypes
 Public
 */
 
-void deletePostfixTree(PostfixTree** root){
+void printPostfixTree(PostfixTree* root){
     if(root){
+        printf("%c", '(');
+        printPostfixTree(root->left);
+        printf(" %s ", root->token);
+        printPostfixTree(root->right);
+        printf("%c", ')');
+    }
+}
+
+void deletePostfixTree(PostfixTree** root){
+    if(root && *root){
         PostfixTree* tree = *root;
         deletePostfixTree(&(tree->left));
         deletePostfixTree(&(tree->right));
